+++ src/org/datafoodconsortium/connector/codegen/typescript/static/src/IConnectorExportObserver.ts
@@ -1,0 +1,5 @@
+import type { Observer } from './observer';
+
+export default interface IConnectorExportObserver extends Observer<string> {
+  next(json: string): void;
+}