+++ src/org/datafoodconsortium/connector/codegen/typescript/static/src/IConnectorExporter.ts
@@ -1,8 +1,11 @@
 import { Semanticable } from "@virtual-assembly/semantizer";
 import IConnectorExporterOptions from "./IConnectorExporterOptions";
+import { Observable, Subscription } from "./observer";
+import IConnectorExportObserver from "./IConnectorExportObserver";
 
-export default interface IConnectorExporter {
+export default interface IConnectorExporter extends Observable<string> {
 
     export(semanticObjets: Array<Semanticable>, options?: IConnectorExporterOptions): Promise<string>;
-    
+
+    subscribe(observer: IConnectorExportObserver): Subscription;
 }