+++ src/org/datafoodconsortium/connector/codegen/typescript/static/test/Address.test.js
@@ -1,6 +1,7 @@
 import expect from 'node:assert';
 import { test } from 'node:test';
 import Connector from "../lib/Connector.js";
+import { assertSemanticEqual, TestObserver } from './utils.js';
 
 const connector = new Connector();
 
@@ -15,15 +16,27 @@ const address = connector.createAddress({
 const json = '{"@context":"https://www.datafoodconsortium.org","@id":"http://myplatform.com/address/address1","@type":"dfc-b:Address","dfc-b:hasCity":"Brussels","dfc-b:hasCountry":"Belgium","dfc-b:hasPostalCode":"00001","dfc-b:hasStreet":"1, place or Europe"}';
 
 test('Address:import', async () => {
+    const testObs = new TestObserver(address, assertSemanticEqual);
+    const testSub = connector.subscribe('import', testObs);
     const imported = await connector.import(json);
     const expected = imported[0];
     expect.strictEqual(imported.length, 1);
     expect.strictEqual(expected.equals(address), true);
+    expect.doesNotThrow(() => {
+        testObs.complete();
+        testSub.unsubscribe();
+    }, '#unsubscribe');
 });
 
 test('Address:export', async () => {
+    const testObs = new TestObserver(json, expect.strictEqual);
+    const testSub = connector.subscribe('export', testObs);
     const serialized = await connector.export([address]);
     expect.strictEqual(serialized, json);
+    expect.doesNotThrow(() => {
+        testObs.complete();
+        testSub.unsubscribe();
+    }, '#unsubscribe');
 });
 
 test('Address:getSemanticId', () => {