+++ src/org/datafoodconsortium/connector/codegen/typescript/static/test/AllergenCharacteristic.test.js
@@ -2,6 +2,8 @@ import * as fs from 'fs';
 import expect from 'node:assert';
 import { test } from 'node:test';
 import Connector from "../lib/Connector.js";
+import { assertSemanticEqual, TestObserver } from './utils.js';
+
 const measures = JSON.parse(fs.readFileSync('./test/thesaurus/measures.json'));
 
 const connector = new Connector();
@@ -19,10 +21,16 @@ const allergenCharacteristic = connector.createAllergenCharacteristic({
 const json = `{"@context":"https://www.datafoodconsortium.org","@id":"_:b1","@type":"dfc-b:AllergenCharacteristic","dfc-b:hasAllergenDimension":"dfc-m:Peanuts","dfc-b:hasUnit":"dfc-m:Kilogram","dfc-b:value":"1"}`;
 
 test('AllergenCharacteristic:import', async () => {
+    const testObs = new TestObserver(allergenCharacteristic, assertSemanticEqual);
+    const testSub = connector.subscribe('import', testObs);
     const imported = await connector.import(json);
     const importedAllergenCharacteristic = imported[0];
     expect.strictEqual(imported.length, 1);
     expect.strictEqual(importedAllergenCharacteristic.equals(allergenCharacteristic), true);
+    expect.doesNotThrow(() => {
+        testObs.complete();
+        testSub.unsubscribe();
+    }, '#unsubscribe');
 });
 
 test('AllergenCharacteristic:export', async () => {