+++ src/org/datafoodconsortium/connector/codegen/typescript/static/test/CustomerCategory.test.js
@@ -1,6 +1,7 @@
 import expect from 'node:assert';
 import { test } from 'node:test';
 import Connector from "../lib/Connector.js";
+import { assertSemanticEqual, TestObserver } from './utils.js';
 
 const connector = new Connector();
 
@@ -12,10 +13,16 @@ const customerCategory = connector.createCustomerCategory({
 const json = `{"@context":"https://www.datafoodconsortium.org","@id":"http://myplatform.com/customerCategory1","@type":"dfc-b:CustomerCategory","dfc-b:description":"description"}`;
 
 test('CustomerCategory:import', async () => {
+    const testObs = new TestObserver(customerCategory, assertSemanticEqual);
+    const testSub = connector.subscribe('import', testObs);
     const imported = await connector.import(json);
     const importedCustomerCategory = imported[0];
     expect.strictEqual(imported.length, 1);
     expect.strictEqual(importedCustomerCategory.equals(customerCategory), true);
+    expect.doesNotThrow(() => {
+        testObs.complete();
+        testSub.unsubscribe();
+    }, '#unsubscribe');
 });
 
 test('CustomerCategory:export', async () => {