+++ src/org/datafoodconsortium/connector/codegen/typescript/static/test/PlannedTransformationLoop.test.js
@@ -2,6 +2,8 @@ import * as fs from 'fs';
 import expect from 'node:assert';
 import { test } from 'node:test';
 import Connector from "../lib/Connector.js";
+import { assertSemanticEqual, TestObserver } from './utils.js';
+
 const measures = JSON.parse(fs.readFileSync('./test/thesaurus/measures.json'));
 const vocabulary = JSON.parse(fs.readFileSync('./test/thesaurus/vocabulary.json'));
 
@@ -53,6 +55,8 @@ const plannedTransformation = connector.createPlannedTransformation({
 const json = '{"@context":"https://www.datafoodconsortium.org","@graph":[{"@id":"_:b1","@type":"dfc-b:QuantitativeValue","dfc-b:hasUnit":"dfc-m:Kilogram","dfc-b:value":"1.2"},{"@id":"_:b2","@type":"dfc-b:QuantitativeValue","dfc-b:hasUnit":"dfc-m:Kilogram","dfc-b:value":"1"},{"@id":"http://myplatform.com/plannedConsumptionFlow","@type":"dfc-b:AsPlannedConsumptionFlow","dfc-b:consumes":"http://myplatform.com/inputProduct","dfc-b:hasQuantity":"_:b1"},{"@id":"http://myplatform.com/plannedProductionFlow","@type":"dfc-b:AsPlannedProductionFlow","dfc-b:hasQuantity":"_:b2","dfc-b:produces":"http://myplatform.com/outputProduct"},{"@id":"http://myplatform.com/transformation","@type":"dfc-b:AsPlannedTransformation","dfc-b:hasIncome":"http://myplatform.com/plannedConsumptionFlow","dfc-b:hasOutcome":"http://myplatform.com/plannedProductionFlow","dfc-b:hasTransformationType":"dfc-v:modify"}]}';
 
 test('PlannedTransformationLoop:import', async () => {
+    const testObs = new TestObserver(plannedConsumptionFlow, assertSemanticEqual);
+    const testSub = connector.subscribe('import', testObs);
     const imported = await connector.import(json);
     const expectedPlannedConsumptionFlow = imported[0];
     const expectedPlannedProductionFlow = imported[1];
@@ -61,6 +65,10 @@ test('PlannedTransformationLoop:import', async () => {
     expect.strictEqual(expectedPlannedTransformation.equals(plannedTransformation), true);
     expect.strictEqual(expectedPlannedConsumptionFlow.equals(plannedConsumptionFlow), true);
     expect.strictEqual(expectedPlannedProductionFlow.equals(plannedProductionFlow), true);
+    expect.doesNotThrow(() => {
+        testObs.complete();
+        testSub.unsubscribe();
+    }, '#unsubscribe');
 });
 
 test('PlannedTransformationLoop:export', async () => {